added code for PSS and updated main function
Added functions to do Portilla-Simoncelli Scrambling (work in progress) and updated main function so that it can now take many optional parameter inputs and so that the wallpaper size is determined by the visual angle and distance in cm from the viewer to the wallpaper. Added some code documentation as well.